+++ ChangeLog.md
@@ -1,6 +1,11 @@
 
 ## Changes ##
 
+### 15.1.4
+ - ENH: Minor changes to support Java 17
+ - BUG: Given monotonic increasing input (which looks like a ZIP) then backing out incorrectly set monotonicIncreasing to false
+ - INT: Make the code and tests more resilient to a Locale with only a language (i.e. no Country)
+
 ### 15.1.3
  - ENH: Minor improvements to TELEPHONE plugin (be more picky to reduce false positives) and accept Contact?Number as a positive header (to detect more valid items)
  - BUG: Examples should be dependent on 15.+ of FTA not 14.+